feat(handlers): refactor and support handlers per repo
