+++ README.md
@@ -39,7 +39,7 @@ To make your GitHub repository accessible to AI assistants via GitMCP, use the f
 Congratulations! The chosen GitHub project is now fully accessible to your AI.
 
 Replace `{owner}` with your GitHub username or organization name and `{repo}` with your repository name. Once configured, your AI assistant can access the project's documentation seamlessly.
-The dynamic endpoint doesn't require a pre-defined repository. When used, your AI assistant can dynamically input any Github repository to enjoy GitMCP's features.
+The dynamic endpoint doesn't require a pre-defined repository. When used, your AI assistant can dynamically input any GitHub repository to enjoy GitMCP's features.
 
 <video src="https://github.com/user-attachments/assets/2c3afaf9-6c08-436e-9efd-db8710554430"></video>
