+++ api/tools/repoHandlers/handlers.ts
@@ -1,0 +1,34 @@
+import { RepoData, UrlType } from "../../../shared/repoData.js";
+import { RepoHandler } from "./RepoHandler.js";
+import { getDefaultRepoHandler } from "./DefaultRepoHandler.js";
+import { getGenericRepoHandler } from "./GenericRepoHandler.js";
+import { getThreejsRepoHandler } from "./ThreejsRepoHandler.js";
+
+const handlers: RepoHandlerMap = {
+  // handle all types of urls for three.js
+  "all::mrdoob/three.js": getThreejsRepoHandler(),
+  // handle only the github type of urls for "generic" repos
+  "github::docs/": getGenericRepoHandler(),
+};
+
+export function getHandlerByRepoData(repoData: RepoData): RepoHandler {
+  const repoKey = `${repoData.owner ?? ""}/${repoData.repo ?? ""}` as RepoKey;
+
+  return (
+    // check if the keyWithUrlType is in the handlers
+    handlers[`${repoData.urlType}::${repoKey}` as UrlTypeRepoKey] ??
+    // check if the allKey is in the handlers
+    handlers[`all::${repoKey}` as AllRepoKey] ??
+    // if not, return the default handler
+    getDefaultRepoHandler()
+  );
+}
+
+type RepoKey = `${string}/${string}`;
+type UrlTypeRepoKey = `${UrlType}::${RepoKey}`;
+type AllRepoKey = `all::${RepoKey}`;
+type MapRepoKey = UrlTypeRepoKey | AllRepoKey;
+
+type RepoHandlerMap = {
+  [key in MapRepoKey]: RepoHandler;
+};
